# express-api/api-spec/halp_api_openapi_3.1.yaml
openapi: "3.1.0"
info:
    title: Halp API
    version: 0.1.0
...
paths:
    /signup:
        post:
            summary: creates a new user
            # add a description with password requirements
            operationId: createUserId
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UserCreateRequestSchema"
            responses:
                "201":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /login:
        post:
            summary: authenticates a user
            description: authenticates the user with the provided password hash and returns an access-token as well as a refresh-token. The access token has an expiration time of 5 minutes, the refresh token has an expiration time of 90 days.
            operationId: loginUserId
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LoginRequestSchema"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccessTokenResponseSchema"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /logout:
        post:
            summary: invalidates access tokens
            operationId: logoutUserId
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/LogoutRequestSchema"
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /refresh-token:
        parameters:
            - name: refreshToken
              in: cookie
              required: true
              schema:
                  type: string
        get:
            # TODO also represent the cookie returned
            summary: refreshes access-token with provided refresh-token
            operationId: refreshTokenId
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/AccessTokenResponseSchema"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"

    # TODO add /reset-password:
...
    /auth/teams:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
        post:
            summary: create team
            operationId: createTeamId
            tags:
                - auth
                - team
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TeamCreateRequestSchema"
            responses:
                "201":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
        get:
            summary: get teams
            operationId: getTeamsId
            tags:
                - auth
                - team
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                contains:
                                    type: object
                                    properties:
                                        teams:
                                            type: array
                                            items:
                                                $ref: "#/components/schemas/TeamEnrichedSchema"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/teams/{teamId}:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/TeamIdParameter"
        delete:
            summary: delete team
            operationId: deleteTeamId
            tags:
                - auth
                - team
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/teams/{teamId}/subscribers:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/TeamIdParameter"
        post:
            summary: subscribe logged in user to team
            operationId: subscribeToTeamId
            tags:
                - auth
                - team
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
        delete:
            summary: unsubscribe logged in user from team
            operationId: unsubscribeFromTeamId
            tags:
                - auth
                - team
            responses:
                "200":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
    /auth/teams/{teamId}/events:
        parameters:
            - $ref: "#/components/parameters/AccessTokenParameter"
            - $ref: "#/components/parameters/TeamIdParameter"
        post:
            summary: create event
            operationId: createEventId
            tags:
                - auth
                - event
                - admin
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/EventCreateRequestSchema"
            responses:
                "201":
                    $ref: "#/components/responses/DefaultSuccessResponse"
                    # TODO replace with actual object
                "default":
                    $ref: "#/components/responses/DefaultErrorResponse"
...
components:
    parameters:
        AccessTokenParameter:
            name: Authorization
            in: header
            required: true
            schema:
                type: string
                # TODO add regex to match 'Bearer .+'
        TeamIdParameter:
            name: teamId
            in: path
            required: true
            schema:
                type: string
...
    responses:
        DefaultSuccessResponse:
            description: OK
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/DefaultSuccessResponseSchema"
        DefaultErrorResponse:
            description: Error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/DefaultErrorResponseSchema"
    schemas:
...
        UserCreateRequestSchema:
            type: object
            properties:
                displayName:
                    type: string
                email:
                    type: string
                password:
                    type: string
                    format: password
            required:
                - displayName
                - email
                - password
        LoginRequestSchema:
            type: object
            properties:
                email:
                    type: string
                password:
                    type: string
                    format: password
            required:
                - email
                - password
        LogoutRequestSchema:
            type: object
            properties:
                email:
                    type: string
            required:
                - email
...
        TeamCreateRequestSchema:
            type: object
            properties:
                teamName:
                    type: string
            required:
                - teamName
        EventCreateRequestSchema:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                dateTime:
                    type: number
                jobs:
                    type: array
                    minItems: 1
                    items:
                        $ref: "#/components/schemas/JobCreatorSchema"
            required:
                - name
                - dateTime
                - jobs
...
        DefaultSuccessResponseSchema:
            type: object
            properties:
                message:
                    type: string
            required:
                - message
        DefaultErrorResponseSchema:
            type: object
            properties:
                errorMessage:
                    type: string
            required:
                - errorMessage
        AccessTokenResponseSchema:
            type: object
            properties:
                accessToken:
                    type: string
            required:
                - accessToken
...
        TeamEnrichedSchema:
            type: object
            properties:
                id:
                    type: string
                adminId:
                    type: string
                name:
                    type: string
                isSubscribed:
                    type: boolean
                isAdmin:
                    type: boolean
            required:
                - id
                - adminId
                - name
                - isSubscribed
                - isAdmin
...
        JobCreatorSchema:
            type: object
            properties:
                jobName:
                    type: string
                count:
                    type: number
                    # TODO positive interger, ...
            required:
                - jobName
                - count
...